# service_database_application/DatabaseServices/bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "DatabaseServices",
    "version": "v1"
  },
  "paths": {
...
    "/api/Patients/{documentID}": {
      "get": {
        "tags": [
          "Patients"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "documentID",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/PatientDTO"
            }
          }
        }
      }
    },
    "/api/Patients/new-patient": {
      "post": {
        "tags": [
          "Patients"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PatientDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Patient"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "BloodGroup": {
      "format": "int32",
      "enum": [
        0,
        1,
        2,
        3
      ],
      "type": "integer"
    },
...
    "Patient": {
      "type": "object",
      "properties": {
        "patientID": {
          "format": "int32",
          "type": "integer"
        },
        "patientInfo": {
          "$ref": "#/definitions/User"
        },
        "bloodgroup": {
          "$ref": "#/definitions/BloodGroup"
        },
        "rh": {
          "$ref": "#/definitions/Rh"
        },
        "sex": {
          "$ref": "#/definitions/Sex"
        },
        "userID": {
          "format": "int32",
          "type": "integer"
        },
        "documentID": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "PatientDTO": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "bloodGroup": {
          "type": "string"
        },
        "rh": {
          "type": "string"
        },
        "sex": {
          "type": "string"
        },
        "documentID": {
          "format": "int32",
          "type": "integer"
        },
        "patientID": {
          "format": "int32",
          "type": "integer"
        },
        "birthDate": {
          "format": "date-time",
          "type": "string"
        },
        "patientAge": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "Rh": {
      "format": "int32",
      "enum": [
        0,
        1,
        2
      ],
      "type": "integer"
    },
    "Sex": {
      "format": "int32",
      "enum": [
        0,
        1,
        2
      ],
      "type": "integer"
    },
    "User": {
      "type": "object",
      "properties": {
        "userID": {
          "format": "int32",
          "type": "integer"
        },
        "firstName": {
          "type": "string"
        },
        "secondName": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "birthDate": {
          "format": "date-time",
          "type": "string"
        }
      }
    }
  }
}